Add option to load a jenv version when invoking MkApk
This commit checks for the JENV_VERSION environment variable. If the
environment variable is set when invoking makeApkWithSources, `jenv
local $JENV_VERSION` will be called. This will set the local java
version to use when running the makeApkWithSources function.

The consequence of this is that the user can specify a non-default java
version to run their tests with.